internal/ci: update CI patterns to match main CUE repo
This brings the CI pattern in this repo in line with that used in the
main CUE repo, as of https://cuelang.org/cl/552206. This change can
broadly be summarised as:

* Copy lots of stuff from the main CUE repo
* Adapt to the patterns it espouses
* Ensure the trybot workflow is ~identical to the current test workflow
* Change "nothing else"

The contents of the base package are a direct copy from the main CUE
repo as of commit 53d3bc715329847fc20b3d7a14ed13606649f0a0. The contents
of this package can effectively be considered a vendor therefore.

The repo package follows the model from the main CUE repo of
establishing repo-specific values. It largely follows defaults, and
embeds base as the main CUE repo does to template.

The trybot pattern from the main CUE repo is followed. The steps are
otherwise untouched, modulo adding the standard template steps for
checking out code, Go caching, and checking the git working tree is
porcelain at the end of the workflow run.

We update the vendor of the GitHub Actions schema to match the vendor
used in the main CUE repo.

We also adopt the workflows pattern from the main CUE repo within the
github package, and embed base.trybotWorkflows as a result.  ci_tool
contains the corresponding changes to work this new pattern.

We switch to using cue v0.5.0-beta.5 to remain consistent with the
version used in the vendor copy from the main CUE repo.

Also generate codereview.cfg rather than maintain by hand.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I2809d03a5bd359d97bd4be63b73e8adbf5e71c76
Reviewed-on: https://review.gerrithub.io/c/cue-lang/proposal/+/552274
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>