internal/ci: re-baseline base from cue-lang/cue@c360ba88
Noticed in passing as a result of a now defunct CL 1200735, that this CL
will supersede.

Includes improvements based on CI setup in cue-lang/vscode-cue, where
the presence of cue in PATH is assumed. The setup-cue action is used
with a "latest" version.

This change moves this repo to be based on the CUE Central Registry for
the curated GitHub actions schema.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I10447b2d13dfdbd1fe067a5462fe59e15b1fae8b
Reviewed-on: https://review.gerrithub.io/c/cue-lang/proposal/+/1204957
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>